refactor(tests): reconcile tests/ui into tests/tests_ui
- Replace auto-spawning server conftest with simple live-server
  standard (localhost:8000, matching tests/ui convention)
- Add ai_providers_page fixture to conftest
- git mv tests/ui/screens/{test_import_article,test_import_hashnode,
  test_settings}.py into tests/tests_ui/screens/
- git mv tests/ui/bridges/test_overview_import_bridge.py into
  tests/tests_ui/bridges/
- git mv tests/ui/outputs/ into tests/tests_ui/outputs/
- Update all moved-file imports from tests.ui.conftest
  to tests.tests_ui.conftest
- Delete tests/ui/ entirely
