+++ tests/tests_ui/bridges/test_overview_import_bridge.py
@@ -10,7 +10,7 @@ Run:
 """
 import pytest
 
-from tests.ui.conftest import BASE_URL
+from tests.tests_ui.conftest import BASE_URL
 
 pytestmark = pytest.mark.browser
